Performance: Move fields to basefields
When we have a very large media library the query we used to search
with was slow and cpu intensive. This is because we had to use left
joins for the description, copyright, width, height and alternate
fields.

This commit further improves performance of the gallery overview
by loading all rendered media items at once.